remove old actions
