+++ src/Components/Loader/types.ts
@@ -1,6 +1,3 @@
-export const LOADING_START = "LOADING_START";
-export const LOADING_END = "LOADING_END";
-export const LOGIN = "LOGIN";
-export const LOGIN_FAILURE = "LOGIN_FAILURE";
-export const LOGOUT = "LOGOUT";
-export const LOGOUT_FAILURE = "LOGOUT_FAILURE";
+export type LoaderState = boolean;
+
+export const loadingDefaultState: LoaderState = true;
