+++ src/Components/NameForm/types.ts
@@ -1,4 +1,22 @@
-export const LOGIN = "LOGIN";
-export const LOGIN_FAILURE = "LOGIN_FAILURE";
-export const LOGOUT = "LOGOUT";
-export const LOGOUT_FAILURE = "LOGOUT_FAILURE";
+enum errorTypes {
+  loginError,
+  logoutError,
+}
+
+type errorLog = Array<keyof typeof errorTypes>;
+
+interface AuthState {
+  name: string;
+  isAuth: boolean;
+  errorLog: errorLog;
+}
+
+export const authDefaultState: AuthState = {
+  name: "",
+  isAuth: false,
+  errorLog: [],
+};
+
+export type payloadType = {
+  name: string;
+};
